requirements.txt
beautifulsoup4==4.8.2
censys==0.0.8
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
colorama==0.4.3
cryptography==2.8
-e git+https://github.com/ArmandSyah/CSI4900-Phishing-Website-Detection.git@f8c642d72ceda511e732872e4afb60a1477c2b1a#egg=CSI4900_phishing
cssselect==1.1.0
docker==4.2.0
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
filetype==1.0.5
fire==0.2.1
html-similarity==0.3.2
idna==2.8
ImageHash==4.0
joblib==0.14.1
lxml==4.5.0
mmh3==2.5.1
netaddr==0.7.19
numpy==1.18.1
pandas==1.0.1
parse==1.14.0
parsel==1.2.0
Pillow==7.0.0
pycparser==2.20
pymongo==3.10.1
pyOpenSSL==19.1.0
pypiwin32==223
pyquery==1.4.1
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
PyWavelets==1.1.1
pywebcopy==6.1.1
pywin32==227
requests==2.22.0
requests-file==1.4.3
rope==0.16.0
scikit-learn==0.22.2.post1
scipy==1.4.1
shodan==1.21.3
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
termcolor==1.1.0
tldextract==2.2.2
urllib3==1.25.8
w3lib==1.21.0
websocket-client==0.57.0
XlsxWriter==1.2.8
